api ok

+++ src/main/java/com/erickfranco/cv_api/services/DiplomeService.java
@@ -1,6 +1,7 @@
 package com.erickfranco.cv_api.services;
 
 
+import com.erickfranco.cv_api.config.exception.BadRequestException;
 import com.erickfranco.cv_api.config.exception.NotFoundException;
 import com.erickfranco.cv_api.models.Diplome;
 import com.erickfranco.cv_api.repositories.DiplomeRepository;
@@ -34,10 +35,17 @@ public class DiplomeService {
 
 
     public Diplome createDiplome(Diplome diplome) {
+        if (diplome.getNomDiplome().isEmpty()) {
+            throw new BadRequestException("Le nom du diplome est obligatoire");
+        }
         return diplomeRepository.save(diplome);
     }
 
+
     public void deleteDiplomeById(Integer id) {
+        if (!diplomeRepository.existsById(id)) {
+            throw new NotFoundException("Le Diplome que vous souhaitez l'eliminer avec l'id " + id + " n'existe pas ");
+        }
         diplomeRepository.deleteById(id);
     }
 }
